analytics/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
mysql-connector-python==8.3.0
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.0
pydantic==2.5.3
scikit-learn==1.4.0
matplotlib==3.8.2
seaborn==0.13.1
